+++ _11114_jilijili_shop/src/main/java/top/jilijili/shop/entity/Categories.java
@@ -1,12 +1,12 @@
 package top.jilijili.shop.entity;
 
-import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
+import lombok.Data;
+
 import java.io.Serializable;
 import java.util.Date;
-import lombok.Data;
 
 /**
  * 商品分类表
@@ -19,7 +19,7 @@ public class Categories implements Serializable {
      * 分类ID
      */
     @TableId(value = "category_id")
-    private Integer categoryId;
+    private Long categoryId;
 
     /**
      * 分类名称
